pyproject.toml
[project]
name = "obsidian-ontology-mcp"
version = "0.2.0"
description = "MCP server for SKOS/Schema.org ontology extraction from Obsidian vaults"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["mcp", "obsidian", "skos", "knowledge-graph", "ontology", "schema.org"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastmcp>=2.0.0",
    "networkx>=3.2",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "watchdog>=4.0.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "slowapi>=0.1.9",
    "python-json-logger>=2.0.0",
    "structlog>=24.1.0",
    "rdflib>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.1",
    "mypy>=1.8",
    "bandit>=1.7",
    "safety>=3.0",
]

igraph = [
    "igraph>=0.11.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/obsidian-ontology-mcp"
Documentation = "https://github.com/yourusername/obsidian-ontology-mcp/tree/main/docs"
Repository = "https://github.com/yourusername/obsidian-ontology-mcp"
Issues = "https://github.com/yourusername/obsidian-ontology-mcp/issues"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
]
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "frontmatter",
    "watchdog.*",
    "slowapi",
    "pythonjsonlogger",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
